cmake可指定socket缓冲区大小, 增加Err_reset错误 (#196)
The default buffer size of a socket should not be hardcoded in the program; it should be configurable. Additionally, in Linux, adjusting kernel parameters can be more effective in improving the performance or throughput of a program in scenarios such as high-latency "long fat networks" or high-concurrency situations, compared to using setsockopt.